# code/backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Reside
  version: 1.0.0
  description: API do Sistema de Gerenciamento de Condomínios Reside
servers:
  - url: https://reside-production.up.railway.app
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadRulesResponse:
      type: object
      properties:
        message:
          type: string
          example: Regras carregadas com sucesso.
      required:
        - message
    ChatRequest:
      type: object
      properties:
        question:
          type: string
          example: Qual é o horário de uso da churrasqueira?
      required:
        - question
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
          example: A churrasqueira pode ser utilizada das 8h às 22h nos finais de semana.
      required:
        - answer
...
paths:
...
  /condos/{condoId}/rules:
    post:
      tags:
        - Condomínios
      summary: Upload das regras do condomínio (somente síndico)
      security:
        - bearerAuth: []
      parameters:
        - name: condoId
          in: path
          description: ID do condomínio
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Regras carregadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadRulesResponse'
  /condos/{condoId}/chat:
    post:
      tags:
        - Condomínios
      summary: Envia pergunta para o chatbot do condomínio (morador)
      security:
        - bearerAuth: []
      parameters:
        - name: condoId
          in: path
          description: ID do condomínio
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Resposta do chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /invite:
    post:
      summary: Envia convites de cadastro para moradores via e-mail
      description: Recebe uma lista de convites com e-mails e apartamentos, enviando um código de convite por e-mail para cada morador.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                condominium_id:
                  type: integer
                  example: 1
                invites:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    properties:
                      emails:
                        type: array
                        items:
                          type: string
                          format: email
                        minItems: 1
                        example: ["morador1@exemplo.com", "morador2@exemplo.com"]
                      apartments:
                        type: array
                        items:
                          type: string
                        minItems: 1
                        example: ["101", "102"]
                    required:
                      - emails
                      - apartments
              required:
                - condominium_id
                - invites
      responses:
        '201':
          description: Convites gerados e e-mails enviados com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    email:
                      type: string
                      format: email
                    apartment:
                      type: string
                    status:
                      type: string
                      example: "enviado"
        '400':
          description: Dados obrigatórios ausentes ou inválidos.
        '500':
          description: Erro interno no servidor.
  /invite/bind:
    post:
      summary: Vincula o morador ao condomínio usando o código recebido por e-mail
      description: Valida o código de convite e cria o cadastro do morador, vinculando ao condomínio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invite_code:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                document:
                  type: string
                contact_phone:
                  type: string
                google_id:
                  type: string
              required:
                - invite_code
                - first_name
                - last_name
                - document
                - contact_phone
                - google_id
      responses:
        '201':
          description: Morador cadastrado e vinculado com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '400':
          description: Código inválido, expirado ou dados ausentes.
        '500':
          description: Erro interno no servidor.
...
  /condos/{condoId}/rules/latest:
    get:
      tags:
        - Condomínios
      summary: Retorna dados do último arquivo de regras
      security:
        - bearerAuth: []
      parameters:
        - name: condoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Nome e data do arquivo
        '404':
          description: Arquivo não encontrado
        '500':
          description: Erro interno do servidor
...